+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 * Fixed making dataset as aminoacid seqs for MEGA format.
 * Fixed making dataset as degenerated seqs for MEGA format.
 * Fixed making dataset as degenerated seqs for TNT format.
+* Fixed making dataset as aa seqs with specified outgroup for TNT format.
 * Raise ValueError when asked to degenerate seqs that will go to partitioning
   based on codon positions.