+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-0.3.8 (2015-10-30)
+0.3.x (2015-10-30)
 ------------------
 * Fixed making dataset as aminoacid seqs for MEGA format.
 * Fixed making dataset as degenerated seqs for MEGA format.